Add files via upload
I made a website addressing issues for women and facts. How it affects women. In the top I put the image I made on photoshop. I made this website because I am passionate about this topic. Women need to be treated differently and there needs to be a change for women to succeed. Life itself is already an obstacle people face so, why as women do we need to face these obstacles.